Improve API error handling and telemetry integration
- Enhanced error handling across various API endpoints to provide more informative responses and improve user experience.
- Updated telemetry tracking to include additional metrics related to API errors, aiding in performance analysis.
- Refactored existing code to ensure consistent error reporting and telemetry actions across the application.
