+++ bitloops/src/cli/root/metadata.rs
@@ -1,0 +1,15 @@
+pub const ROOT_NAME: &str = "bitloops";
+pub const ROOT_SHORT_ABOUT: &str = "Bitloops CLI";
+pub const ROOT_LONG_ABOUT: &str = r#"The command-line interface for Bitloops
+
+Getting Started:
+  To get started with Bitloops CLI, run 'bitloops start' to launch the
+  daemon, then run 'bitloops init' inside a repository or subproject.
+  For more information, visit:
+  https://docs.bitloops.io/introduction
+
+Environment Variables:
+  ACCESSIBLE    Set to any value (e.g., ACCESSIBLE=1) to enable accessibility
+                mode. This uses simpler text prompts instead of interactive
+                TUI elements, which works better with screen readers.
+"#;
